requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
libretranslatepy==2.1.1
lxml==5.0.1
MarkupSafe==2.1.3
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
requests==2.31.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
